+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 - refactoring (interface manager)
 - adding formatter widgets, merge formatter options with column definition formatter options
 - adding advancved formatter (formatter widget per column)
+- refactoring nunjucks template loader and template interface, now we can add custom templates